aiju_weixin/requirements.txt
Cheetah==2.4.1
Flask==0.10.1
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
MySQL-python==1.2.5
PIL==1.1.6
PyYAML==3.10
Pygments==1.4
SQLAlchemy==0.9.8
Werkzeug==0.9.6
argparse==1.2.1
aws-cfn-bootstrap==1.4
awscli==1.7.0
backports.ssl-match-hostname==3.4.0.2
bcdoc==0.12.2
boto==2.35.1
botocore==0.81.0
chardet==2.0.1
cloud-init==0.7.2
colorama==0.2.5
configobj==4.7.2
docutils==0.11
flufl.enum==4.0.1
iniparse==0.3.1
itsdangerous==0.24
jmespath==0.5.0
kitchen==1.1.1
lockfile==0.8
ordereddict==1.1
paramiko==1.7.5
ply==3.4
pycrypto==2.6.1
pycurl==7.19.0
pygpgme==0.1
pyliblzma==0.5.3
pystache==0.5.3
python-daemon==1.5.2
python-dateutil==2.1
pyxattr==0.5.0
requests==1.2.3
rsa==3.1.2
simplejson==3.5.3
six==1.8.0
urlgrabber==3.9.1
urllib3==1.8.2
virtualenv==1.11.6
wsgiref==0.1.2
yum-metadata-parser==1.1.4
